+++ errors.go
@@ -8,6 +8,6 @@ var (
 	// ErrInput is returned when an input is invalid.
 	ErrInput = errors.New("invalid input")
 
-	// ErrClosed is returned when the Exchange is closed and is being re-used.
+	// ErrClosed is returned when the Client is closed and is being reused.
 	ErrClosed = errors.New("exchange is already closed")
 )