+++ CHANGELOG.md
@@ -5,6 +5,16 @@
 
 
 
+<a name="v0.1.1"></a>
+## [v0.1.1] - 2020-07-24
+
+### Added
+- Add queue implementation and documentations
+
+### Implement
+- Implement publishing and consuming of messages
+
+
 <a name="v0.1.0"></a>
 ## v0.1.0 - 2020-07-23
 
@@ -13,4 +23,5 @@
 
 
 
-[Unreleased]: https://github.com/blokur/harego/compare/v0.1.0...HEAD
+[Unreleased]: https://github.com/blokur/harego/compare/v0.1.1...HEAD
+[v0.1.1]: https://github.com/blokur/harego/compare/v0.1.0...v0.1.1